refactor(chess): ChessService, ChessRepository 인터페이스 삭제
